refactor: SegmentedControl 관련 사항 변경
- 각각의 layout이 collectionView하나씩을 가지게 변경
- toggle() 과 becomeFirstResponder() 를 사용하여 오류 방지

+++ OpenMarket/OpenMarket/View/Extension/String+Extension.swift
@@ -32,7 +32,7 @@ extension String {
 
 extension String {
     var decimal: String? {
-        guard let number = Int(self) else {
+        guard let number = Double(self) else {
             return nil
         }
         let numberformatter = NumberFormatter()
